Cast the lm_head weight to the compute dtype in the chunked projections
Under FSDP2 mixed precision the patched forward reads the fp32 sharded master
weight directly while the hidden states come out bf16, so the tensor-core
projection failed with a dtype mismatch (tests/distributed test_sft_peft[fsdp2]).
Cast the weight to the hidden-states dtype, which is what lm_head's own forward
computes in. Same fix in all three copies: SFT, distillation, async distillation.

+++ trl/experimental/async_distillation/async_distillation_trainer.py
@@ -225,8 +225,9 @@ def _jsd_loss_chunk(
         are `torch.no_grad()` sums for this chunk only — callers accumulate them across chunks and reduce across ranks,
         exactly as the non-chunked path already did.
     """
-    # Project in the model dtype and upcast only afterwards, as the other chunked projections do.
-    logits = (hidden_chunk @ lm_head_weight.t()).float()
+    # Project in the compute dtype and upcast only afterwards, as the other chunked projections do, casting
+    # the weight to the hidden-states dtype in case it is an fp32 master weight (FSDP2 mixed precision).
+    logits = (hidden_chunk @ lm_head_weight.to(hidden_chunk.dtype).t()).float()
     if lm_head_bias is not None:
         logits = logits + lm_head_bias.float()
     if logit_scale != 1.0:
